+++ src/utils.rs
@@ -414,7 +414,7 @@ async fn update_query_database(
             // Has to be updated over all auctions instead of comparing previous lowest bins with new auctions
             if update_lowestbin && ele.get("bin") {
                 let internal_id: String = ele.get("internal_id");
-                let lowestbin_price: f64 = ele.get("lowestbin_price");
+                let lowestbin_price: f32 = ele.get("lowestbin_price");
                 update_lower_else_insert(&internal_id, lowestbin_price, bin_prices);
             }
         }
